Moved LCurve to new module "hyper_param"
